refactor(agent): move agent catalog to shared for renderer access
Move AGENT_CATALOG and helpers from src/main/services/agents/agent-catalog.ts
to src/shared/agent-catalog.ts so renderer code (settings page) can import
display fields without violating the main⊥renderer depcruise boundary.
Update detection-service, agent-launch, and their tests to import from the
new shared location directly (avoids barrel file lint error noBarrelFile).
