+++ src/shared/agent-catalog.ts
